+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - Provided `README.md` with basic project information and setup instructions.
 ## [Unreleased]
 ### Added
-- Added `power(a,b)` function in `functions.py` to perform division.
+- Added `power(a,b)` function in `functions.py` to perform power.
 ### Fixed
 - power 함수에 내용을 추가했습니다. 
+### Added
+- Added `multiple(a,b)` function in `functions.py` to perform multiple.